requirements.txt
pymongo
mongoengine
bcrypt
Flask
flask_bcrypt
PyJWT
cffi
click
colorama
dnspython
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
pycparser
Werkzeug
zipp
flask_pymongo

# pip install -r requirements.txt